refactor: Optimize colocation structure - move page-specific features closer to usage
- Move service-list.tsx to app/components/ (top page only)
- Move contact libs to app/contact/libs/ (contact page only)
- Move contact types to app/contact/types/ (contact page only)
- Update all import statements to use relative paths
- Remove contact exports from global libs and types
- Clean up empty directories
- Build passing ✓

Perfect colocation achieved:
- app/components/service-list.tsx (top page specific)
- app/contact/libs/ (contact API functions)
- app/contact/types/ (contact type definitions)

This completes the true colocation principle where features are placed as close as possible to their usage.
