+++ src/libs/index.ts
@@ -1,2 +1,1 @@
 export * from './fetch'
-export * from './contact'
