+++ src/types/index.ts
@@ -1,3 +1,2 @@
-export * from './contact'
 export * from './feed'
 export * from './microcms'
